+++ src/spatialdata/_io/_utils.py
@@ -336,36 +336,6 @@ def _is_element_self_contained(
     return all(_backed_elements_contained_in_path(path=element_path, object=element))
 
 
-@singledispatch
-def get_channels(data: Any) -> list[Any]:
-    """Get channels from data.
-
-    Parameters
-    ----------
-    data
-        data to get channels from
-
-    Returns
-    -------
-    List of channels
-    """
-    raise ValueError(f"Cannot get channels from {type(data)}")
-
-
-@get_channels.register
-def _(data: SpatialImage) -> list[Any]:
-    return data.coords["c"].values.tolist()  # type: ignore[no-any-return]
-
-
-@get_channels.register
-def _(data: MultiscaleSpatialImage) -> list[Any]:
-    name = list({list(data[i].data_vars.keys())[0] for i in data})[0]
-    channels = {tuple(data[i][name].coords["c"].values) for i in data}
-    if len(channels) > 1:
-        raise ValueError(f"Channels are not consistent across scales: {channels}")
-    return list(next(iter(channels)))
-
-
 def save_transformations(sdata: SpatialData) -> None:
     """
     Save all the transformations of a SpatialData object to disk.
